Do some small-constant-factor space optimizations.
To store 47K traces we are now at 18MiB. Close to 0.4M per thousand
traces. Before it was a little closer to 1M per thousand. This is a
little better.

And note that it is a LOT better that what I reported just a few days
ago (1.6G for 27K traces). This is like 100x more compact -- partly, of
course, because it simply stores less data (rank order instead of
physical time).