Fail fast when the output stream doesn't close.
IDK why this happens exactly but in this case it happened because the
binaries for the federates simply weren't there for some reason --
accidentally deleted?

Also, I notice that polling less frequently does not actually increase
the number of tests that we can run. Polling each second drops it in the
range of 700-800; dropping the period to 10 milliseconds gets the
performance that we had before the change (this is runs per 20 seconds,
except that we do not kill runs that started before the 20 seconds were
up -- those count too, so it is a very complicated metric). More
experimentation with this should be done. It might still be a good idea
but I spoke too soon when I declared a "clear win."