Optimize and tinker with parameters.
A search radius of 32 gives visibly smaller before_and_after sets than
does a search radius of 128, perhaps by a factor of 2 in a few cases.
But it gives larger sets than a radius of 8 or 16 (although those might
be acceptable), at the cost of a noticeably higher end-to-end latency
(on the order of 2x or more perhaps -- as you would expect, accounting
for more cycles as well as more memory access). For heuristic reasons I
estimate this to be a reasonable tradeoff.