Fix a serialization/deserialization bug.
The problem was that the saved_up_to fields were not being updated
whenever we do a save. This is really a symptom of my not obeying what
the serialization and deserialization methods are supposed to do --
perhaps I should not have tried to hook them into the serde framework in
this way and instead just pushed the serde api calls deeper into the
system with the top level just being an abstract "save to file" call.

This also de-optimizes some stuff in the name of debuggability. I should
really create tests to check for regressions here and re-optimize the
stuff.