Parallelism-related changes.
Notably, this includes parallelizing the streaming transpositions
algorithm in a way that may affect accuracy (although it is not obvious
whether it would make accuracy better or worse). This commit includes
some commented-out code. If I use that code instead of par_record_all,
then I get very weird output from the viz (the chart is all colored red
and the histograms are not rendered??? and it takes 3 minutes to run,
which is way out of proportion to the 6 seconds when parallelized -- I
only have 6 or 12 cores depending how you count -- with hammering all on
one core for most of the time).