Update according to changes in reactor-c.
This commit also includes a fix for a serious bug in the way trace files
were being interpreted. Previously I assumed that they were in
chronological order, but this is not true, so I need to read them
eagerly in their entirety (they are not usually that large, so this is
OK) and sort all the records by physical time.