Simplify streaming-transpositions.
Now I feel pretty silly for spending so much time on it yesterday now
that it seems clear that the problem is so trivial. This change was
motivated by an OOM on Wessel (which has 90+ GB RAM!). The "weird
behavior" that I was also seeing locally in relation to this algorithm
is also probably OOM. I think this commit should fix that.