More bugfixes related to streaming-transpositions.
The real reason for the "weird behavior" was that the plotting framework
was getting input that it didn't expect -- basically garbage input.
Probably the plot looked weird (all red) because there was a red plot
line that was jumping up and down with high frequency. The plotting
framework is the reason why it was previously hammering on a single core
for minutes at a time.