Hack together a fix for c-blocking-server-complex.
This solves the problem that was preventing sockets from being recovered
and reused in multiple runs. It breaks things for the TCP variant of the
blocking server library, and it commits a lot of SWE sins, and some
examples don't work anymore, but I'm in a hurry and this fixes the
feature that I need to work right now.