Fix a somewhat rare deadlock.
The deadlock only happens when the connection-granting tokio task gets
in an invalid state, waiting forever for a frame forever after a new
connection has already been requested and it is time to move on.
Probably the invalid state is due to clients misbehaving, which is
"normal" because clients are by assumption unreliable (they are part of
the SUT).