Make the way measurements are done fairer.
I could be accused of changing the way measurements are done in order to
outperform the baseline. But truly, the baseline was being run 5x longer
and that's not a fair comparison. Even after I do this it won't be a
fair comparison because the non-baseline has a stopping condition that
can make it finish faster. In the report I will acknowledge that in the
limit if they are run long enough the two perform the same.