More work on the upper-bounding project.
I think that this strategy of pseudo-events is going to get unmanageable
to express syntactically by creating a new enum element for each
possible thing without incorporating the relation concept to qualify it.
We have disjunctions of events, we have "first"s, and now it looks like
we will need to represent the last event -- and not only that, but the
last event that matches a certain predicate.