Factor out reusable predicate-building.
By predicate-building I mean connectives like and, or, etc. Only the
atoms should be customized. This will make it easier to add more
connectives such as implies and exactly-one (AKA n-way XOR), and it will
make it easier to reuse code that uses those connectives, such as code
that ad-hoc identifies particular predicates as being trivial (i.e.,
constant functions) or else as being NOT the simplest elements of their
equivalence classes wrt the operation of projecting to boolean
functions.